make a hook for compiling
